Got the xpconfig.pl script working. Here is what I had to change if you are curious.
Original (Does not work with 9.2.1)
Line 99
$vlan_cmd = "ENT-VLAN::VLAN-$vlan:1:::NAME=VLAN$vlan";

Modified (Working)
Line 99
$vlan_cmd = "ENT-VLAN::VLAN-$opt_s-$opt_l-$vlan:1:::NAME=VLAN$vlan";

Makes sense. The node needs to know what shelf/slot to add the S-VLAN to. Not sure why the documentation leaves that out of the examples as the command will not work without it.

Anyways<85>working fine now. That<92>s going to save me a ton of time.

Attached is the modified working version.

Take care.

Tim Calvin
Texas Lone Star Network